+++ src/features/utils/instrument-base.js
@@ -90,18 +90,18 @@ export class InstrumentBase extends FeatureBase {
         if (this.featureName === FEATURE_NAMES.sessionReplay) this.abortHandler?.() // SR should stop recording if session DNE
       }
 
-      // Create a single Aggregator for this agent if DNE yet; to be used by jserror endpoint features.
-      if (!agentRef.sharedAggregator) {
-        agentRef.sharedAggregator = import(/* webpackChunkName: "shared-aggregator" */ '../../common/aggregate/event-aggregator')
-        const { EventAggregator } = await agentRef.sharedAggregator
-        agentRef.sharedAggregator = new EventAggregator()
-      } else await agentRef.sharedAggregator // if another feature is already importing the aggregator, wait for it to finish
-
       /**
        * Note this try-catch differs from the one in Agent.run() in that it's placed later in a page's lifecycle and
        * it's only responsible for aborting its one specific feature, rather than all.
        */
       try {
+        // Create a single Aggregator for this agent if DNE yet; to be used by jserror endpoint features.
+        if (!agentRef.sharedAggregator) {
+          agentRef.sharedAggregator = import(/* webpackChunkName: "shared-aggregator" */ '../../common/aggregate/event-aggregator')
+          const { EventAggregator } = await agentRef.sharedAggregator
+          agentRef.sharedAggregator = new EventAggregator()
+        } else await agentRef.sharedAggregator // if another feature is already importing the aggregator, wait for it to finish
+
         if (!this.#shouldImportAgg(this.featureName, session)) {
           drain(this.agentIdentifier, this.featureName)
           loadedSuccessfully(false) // aggregate module isn't loaded at all